init: decomposition app on auth core shared youtube modules for Routing task
