+++ src/app/core/components/header/header-logo/header-logo.component.scss
