+++ src/app/core/models/search-item.model.ts
