+++ src/app/core/models/search-response.model.ts
