+++ src/app/youtube/components/search-result/search-result-item/search-result-item.component.scss
