+++ src/app/youtube/components/search-result/search-result.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
-import { GetResponseService } from 'src/app/services/get-response-service.service';
-import { SearchItem } from '../../models/search-item.model';
+import { GetResponseService } from 'src/app/core/services/get-response-service.service';
+import { SearchItem } from '../../../core/models/search-item.model';
 
 @Component({
   selector: 'app-search-result',
