+++ src/app/youtube/directives/style-item-date.directive.ts
